Create generalized, bi-directional relationships
Real life family trees often do not always follow conventional
relationship rules and therefore should be generalized. Many of the
standard relationships (e.g. uncle, aunt, grandparent) can be
extrapolated based on some basic relationships (i.e. parent/child), but
these are not the only possible relationships that should be
considered. Godparent/godchild relationships do not lend themselves to
pedigree tracking, but are an important aspect of a family's history.
Conversely, adoption may or may not be an important aspect of a family
history, but may present interesting or vital insight into the pedigree
or medical history of the family.

The current implementation is a simple bi-directional representation in
which each member has a specific role (e.g. mother/son).

TODO: Several topics may be interesting to explore in the future,
including, but not limited to:

* Vertical versus horizontal relationships: Family members that share
both a vertical (e.g. mother/son) and a horizontal (e.g. husband/wife)
relationship may pose problems in representing a family tree visually,
but are relationships that may exist or have existed in a family tree.
Currently the database assumes that each couple will have exactly one
relationship, therefore representing multiple relationships may
require some redesign.